// packages/wizzi.hub.backend/src/features/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi.plugins\packages\wizzi.plugin.ts\lib\artifacts\ts\module\gen\main.js
    package: @wizzi/plugin.ts@
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.apps\packages\wizzi.hub.backend\.wizzi-override\src\features\types.ts.ittf
    utc time: Wed, 31 Jul 2024 13:44:15 GMT
*/

export type ValidateResult = { 
    isValid: boolean;
    message?: string;
};
;

export type CRUDResult = { 
    oper: string;
    ok: boolean;
    item?: any;
    message?: string;
};
;